ready to the second train
