+++ sh_scripts/install_gcloud.sh
@@ -1,0 +1,4 @@
+#! /bin/bash
+curl https://sdk.cloud.google.com | bash
+
+gcloud init
