+++ tfmodules/path_manager.py
@@ -49,7 +49,7 @@ TF_CNN_TEST_DIR         = TF_CNN_MODULE_DIR      + '/testcodes'
 DATASET_DIR                 = PROJ_HOME     + '/dataset'
 TFRECORD_REALSET_DIR        = DATASET_DIR   + '/tfrecords/realdataset/'
 TFRECORD_TESTSET_DIR        = DATASET_DIR   + '/tfrecords/testdataset/'
-TFRECORD_TESTIMAGE_DIR           = DATASET_DIR   + '/tfrecords/testimagedataset'
+TFRECORD_TESTIMAGE_DIR      = DATASET_DIR   + '/tfrecords/testimagedataset'
 # TRAIN_DATASET_DIR   = DATASET_DIR   + '/train/'
 # TEST_DATASET_SIR    = DATASET_DIR   + '/eval/'
